refactor(types): migrate from tsd to tstyche and others updated
